requirements.txt
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
imageio==2.34.1
importlib_resources==6.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
piq==0.8.0
progressbar2==4.4.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-utils==3.8.2
six==1.16.0
sympy==1.12
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0
zipp==3.18.1
